requirements.txt
google-auth==2.29.0
google-api-python-client==2.127.0
requests==2.31.0
